move remote app view to linked_domain app
