+++ corehq/apps/linked_domain/util.py
@@ -12,3 +12,14 @@ def _clean_json(doc):
         if isinstance(val, list):
             map(_clean_json, val)
     return doc
+
+
+def convert_app_for_remote_linking(latest_master_build):
+    _attachments = latest_master_build.get_attachments()
+    source = latest_master_build.to_json()
+    source['_LAZY_ATTACHMENTS'] = {
+        name: {'content': content}
+        for name, content in _attachments.items()
+    }
+    source.pop("external_blobs", None)
+    return source
